candidate labor representative and proprietor updates

+++ app/Http/Controllers/Candidate/Agency/ProprietorController.php
@@ -1,57 +1,60 @@
 <?php
 
-namespace App\Traits;
+namespace App\Http\Controllers\Candidate\Agency;
 
-use Illuminate\Http\JsonResponse;
+use App\Http\Controllers\Candidate\BackendBaseController;
+use App\Http\Requests\Backend\ProprietorRequest;
+use App\Models\Backend\AuthorizedAgency;
+use App\Models\Backend\Proprietor;
+use App\Traits\ControllerOps;
 use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Contracts\View\Factory;
 use Illuminate\Contracts\View\View;
+use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Session;
 
-trait ControllerOps {
 
-    /**
-     * @return Application|Factory|View|\Illuminate\Foundation\Application
-     */
+class ProprietorController extends BackendBaseController
+{
+    use ControllerOps;
+    protected string $module        = 'candidate.';
+    protected string $base_route    = 'candidate.authorized_agency.proprietor.';
+    protected string $view_path     = 'candidate.authorized_agency.proprietor.';
+    protected string $page          = 'Proprietor';
+    protected string $folder_name   = 'proprietor';
+    protected string $page_title, $page_method, $image_path, $file_path;
+    protected object $model;
 
-    public function index()
+
+    public function __construct()
     {
-        $this->page_method = 'index';
-        $this->page_title  = 'List '.$this->page;
-        $data              = $this->getData();
-        $data['row']       = $this->model->descending()->get();
-
-        return view($this->loadResource($this->view_path.'index'), compact('data'));
+        $this->model            = new Proprietor();
     }
 
     public function getData(): array
     {
-        return [];
-    }
+        $data['agencies'] = AuthorizedAgency::active()->descending()->pluck('title','id');
 
-    public function create()
-    {
-        $this->page_method = 'create';
-        $this->page_title  = 'Create '.$this->page;
-        $data              = $this->getData();
-
-        return view($this->loadResource($this->view_path.'create'), compact('data'));
+        return $data;
     }
 
 
-    public function store(Request $request)
+    /**
+     * Store a newly created resource in storage.
+     *
+     * @param ProprietorRequest $request
+     * @return JsonResponse
+     */
+    public function store(ProprietorRequest $request)
     {
         DB::beginTransaction();
         try {
-            if($request->hasFile('image_input')){
-                $image_name = $this->uploadImage($request->file('image_input'));
-                $request->request->add(['image'=>$image_name]);
-            }
             $request->request->add(['created_by' => auth()->user()->id ]);
 
-            $this->model->create($request->all());
+            $proprietor = $this->model->create($request->all());
+
             Session::flash('success',$this->page.' was created successfully');
             DB::commit();
         } catch (\Exception $e) {
@@ -60,23 +63,6 @@ trait ControllerOps {
         }
 
         return response()->json(route($this->base_route.'index'));
-    }
-
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return Application|Factory|\Illuminate\Foundation\Application|View
-     */
-
-    public function show($id)
-    {
-        $this->page_method = 'show';
-        $this->page_title  = 'Show '.$this->page;
-        $data              = $this->getData();
-        $data['row']       = $this->model->find($id);
-
-        return view($this->loadResource($this->view_path.'show'), compact('data'));
     }
 
     /**
@@ -92,21 +78,28 @@ trait ControllerOps {
         $data              = $this->getData();
         $data['row']       = $this->model->find($id);
 
+        if ($this->checkAuthorizedAgency($data['row']->authorized_agency_id )){
+            return redirect()->back();
+        }
+
         return view($this->loadResource($this->view_path.'edit'), compact('data'));
     }
 
-    public function update(Request $request, $id)
+    /**
+     * Update the specified resource in storage.
+     *
+     * @param ProprietorRequest $request
+     * @param int $id
+     * @return JsonResponse
+     */
+    public function update(ProprietorRequest $request, $id)
     {
         $data['row']       = $this->model->find($id);
 
         DB::beginTransaction();
         try {
-            if($request->hasFile('image_input')){
-                $image_name = $this->updateImage($request->file('image_input'),$data['row']->image);
-                $request->request->add(['image'=>$image_name]);
-            }
-
             $request->request->add(['updated_by' => auth()->user()->id ]);
+
             $data['row']->update($request->all());
 
             Session::flash('success',$this->page.' was updated successfully');
@@ -119,8 +112,6 @@ trait ControllerOps {
         return response()->json(route($this->base_route.'index'));
     }
 
-
-
     /**
      * Remove the specified resource from storage.
      *
@@ -129,6 +120,10 @@ trait ControllerOps {
      */
     public function destroy($id)
     {
+        $data['row'] = $this->model->find($id);
+        if ($this->checkAuthorizedAgency($data['row']->authorized_agency_id )){
+            return redirect()->back();
+        }
         try {
             DB::beginTransaction();
             $this->model->find($id)->forceDelete();
@@ -146,37 +141,14 @@ trait ControllerOps {
     }
 
 
-
-    public function trash()
-    {
-        $this->page_method = 'trash';
-        $this->page_title  = 'Trash '.$this->page;
-        $data              = [];
-        $data['users']     = $this->model->onlyTrashed()->get();
-
-        return view($this->loadResource($this->view_path.'trash'), compact('data'));
-    }
-
-    public function restore(Request $request, $id)
-    {
-
-        DB::beginTransaction();
-        try {
-            $this->model->withTrashed()->find($id)->restore();
-
-            Session::flash('success',$this->page.' restored successfully');
-            DB::commit();
-        } catch (\Exception $e) {
-            DB::rollback();
-            Session::flash('error',$this->page.' restored failed. Something went wrong.');
-        }
-
-        return redirect()->route($this->base_route.'index');
-    }
-
     public function removeTrash(Request $request, $id)
     {
         $data['row']       = $this->model->withTrashed()->find($id);
+
+        if ($this->checkAuthorizedAgency($data['row']->authorized_agency_id )){
+            return redirect()->back();
+        }
+
         DB::beginTransaction();
         try {
             if ($data['row']->image){
@@ -193,6 +165,5 @@ trait ControllerOps {
 
         return redirect()->route($this->base_route.'trash');
     }
-
 
 }
